Fix CI: clang-tidy cast, Bazel missing texture_2d_array, WebGPU paintParams, clang-format
Four independent CI failures from this branch:

- terrain_layer_tweaker.cpp:42 had the same redundant static_cast<double>
  around TransformState::getZoom() (already double) that 7de836b fixed in
  render_terrain.cpp; it fails the Linux clang-tidy build (-Werror). This was
  the cause of all three linux-build-and-test job failures.

- linux-coverage (Bazel) could not compile src/mbgl/gl/drawable_gl.cpp:
  "mbgl/gl/texture_2d_array.hpp: No such file or directory". The GL
  texture_2d_array.{cpp,hpp} added by the terrain depth-instancing port were
  never added to bazel/core.bzl, so only the CMake builds saw them.

- macOS WebGPU (wgpu) failed shader creation with "no definition in scope for
  identifier: paintParams". Both background WGSL vertex shaders call
  apply_drape_transform(..., paintParams.drape_tile) but neither declared the
  GlobalPaintParamsUBO struct or its @group(0) @binding(0) binding - only the
  pattern *fragment* shader did. Added the struct + binding to both vertex
  modules and completed the fragment's stubbed struct so the layouts agree.

- android-build (opengl/vulkan) and pre-commit.ci failed on clang-format;
  ran the hook over the tree.

+++ src/mbgl/renderer/layers/terrain_layer_tweaker.cpp
@@ -39,7 +39,7 @@ void TerrainLayerTweaker::execute(LayerGroupBase& layerGroup, const PaintParamet
     // this many metres into a curtain that hides the cracks between neighbouring
     // tiles at different zoom levels. ~1/5 of the tile's world width at this zoom,
     // matching maplibre-gl-js Terrain.getSkirtLength().
-    const auto zoom = std::max(static_cast<double>(parameters.state.getZoom()), 0.0);
+    const auto zoom = std::max(parameters.state.getZoom(), 0.0);
     const float elevationOffset = static_cast<float>(util::M2PI * util::EARTH_RADIUS_M / std::pow(2.0, zoom) / 5.0);
 
     // Populate layer-level UBO with terrain properties
